fix: region bug + clean up
