+++ docker-compose.yaml
@@ -50,7 +50,7 @@ services:
   liquibase:
     container_name: liquibase_update
     image: liquibase/liquibase
-#    restart: on-failure
+    restart: on-failure
     volumes:
       - ./liquibase:/liquibase/changelog
     command:
