+++ liquibase/changelog.xml
@@ -23,7 +23,7 @@
     <include file="changes/feat_360.sql" relativeToChangelogFile="true"/>
     <include file="changes/feat_389.sql" relativeToChangelogFile="true"/>
     <include file="changes/feat_533.sql" relativeToChangelogFile="true"/>
-    <include file="changes/feat_618.sql" relativeToChangelogFile="true"/>
     <include file="changes/feat_611.sql" relativeToChangelogFile="true"/>
+    <include file="changes/feat_618.sql" relativeToChangelogFile="true"/>
     <include file="changes/feat_618_2.sql" relativeToChangelogFile="true"/>
 </databaseChangeLog>
